Add upload image feature, modal delete button, purchased button

+++ client/src/components/item_modal.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
+//import { Link } from "react-router-dom";
 import Modal from "react-modal";
 import axios from "axios";
 import useItems from "../hooks/useItems";
@@ -18,10 +18,11 @@ const customStyles = {
 Modal.setAppElement("body");
 
 const ItemModal = ({ item, onClose, state }: any) => {
-  let [currentItem, setCurrentItem] = useState(state.curItem);
+  //let [currentItem, setCurrentItem] = useState(state.curItem);
   const [modalIsOpen, setIsOpen] = useState(false);
   const { purchaseItem } = useItems();
   const { deleteItem } = useItems();
+  const { deleteList } = useItems();
   const openModal = () => {
     setIsOpen(true);
   };
@@ -55,6 +56,7 @@ const ItemModal = ({ item, onClose, state }: any) => {
       date: item.date,
       worry: item.worry,
       category_id: item.category_id,
+      image: item.image,
     };
     purchaseItem(item.id);
 
@@ -71,10 +73,8 @@ const ItemModal = ({ item, onClose, state }: any) => {
       }
     }
   };
-  // todo 수정하기 : 어려울듯 .. 제일 마지막에
-  const editItem = async () => {};
 
-  // todo 삭제하기 : 상태변화
+  // * 삭제하기 : 상태변화
   const handleDelete = async () => {
     if (window.confirm("해당 아이템을 삭제하시겠습니까?")) {
       let url = `http://18.217.232.233:8080/item?item_id=${item.id}`;
@@ -85,7 +85,12 @@ const ItemModal = ({ item, onClose, state }: any) => {
         });
         if (res.status === 202) {
           console.log("delete item");
-          deleteItem(item.id);
+          if (
+            item.date.slice(5, 7) === String("0" + (new Date().getMonth() + 1))
+          ) {
+            deleteItem(item.id);
+          }
+          deleteList(item.id);
           onClose({ ...state, curItem: {} });
           setIsOpen(false);
         }
@@ -97,6 +102,8 @@ const ItemModal = ({ item, onClose, state }: any) => {
     }
   };
 
+  // todo 수정하기 : 어려울듯 .. 제일 마지막에
+  console.log(item);
   return (
     <div>
       <Modal
@@ -106,12 +113,32 @@ const ItemModal = ({ item, onClose, state }: any) => {
         style={customStyles}
         contentLabel="Example Modal"
       >
-        <h2>{item.item_name}</h2>
-        <img src="https://res.cloudinary.com/dgggcrkxq/image/upload/v1579705687/noticon/k1amz2nqzg6pgowfuchn.gif" />
-        <img src="https://res.cloudinary.com/dgggcrkxq/image/upload/v1567062612/noticon/fqdjmxuq27tt7o4umaoy.gif" />
-        <div>{item.memo}</div>
-        <div>{item.worry}</div>
-        <div>{`${onFormat(Number(item.item_price))}원`}</div>
+        <h2>
+          <input></input>
+        </h2>
+        <div>
+          <p>이미지</p>
+          <img
+            style={{ width: "300px" }}
+            src="https://asmallgoodthing.s3.ap-northeast-2.amazonaws.com/image/1590626610724.png"
+          />
+        </div>
+        <div>
+          <p>
+            <input></input>
+          </p>
+          {item.date}
+        </div>
+        <div>
+          <p>고민하는 이유</p>
+          {item.memo}
+        </div>
+        <div>
+          <p>고민의 정도</p>
+          {item.worry}
+        </div>
+        <p>금액</p>
+        <div>{`${onFormat(Number(item.item_price))}원`}</div>;
         <form>
           <input />
         </form>
